게시물에 조회수/좋아요/스크랩수 저장, 대댓글 등록/삭제

+++ src/main/java/com/hanaro/wouldyouhana/dto/LikesScrapResponseDTO.java
@@ -17,8 +17,8 @@ public class LikesScrapResponseDTO {
     // 화면에 보여지는 내용
     private String categoryName;
     private String questionTitle;
-    private Integer likeCount;
-    private Integer viewCount;
+    private Long likeCount;
+    private Long viewCount;
     private LocalDateTime createdAt;
     private String bankerName;
 }
